requirements.txt
tqdm
signalrcore
pillow
pandas
seaborn
psutil
ipython
pyyaml
scipy
matplotlib
numpy
--extra-index-url https://download.pytorch.org/whl/cpu 
torch==1.12.0
torchaudio==0.12.0
torchvision==0.13.0
opencv-python-headless
uuid6 
tensorflow-cpu==2.8.2
timm
peewee
fiftyone
fastapi==0.75.2
loguru